Extract responsibility of formatting messages from TextUI to new Formatter class
